fixed shader loader NOTE: do not revert to this commit, half done changes exists!

+++ src/render.cpp
@@ -15,11 +15,11 @@ render::initializeBuffer()
     glBindBuffer(GL_ARRAY_BUFFER, VBO);
 }
 
-
+#include <iostream>
 void
 render::setBuffer(std::vector<float> vertice)
 {
-    glBufferData(GL_ARRAY_BUFFER, static_cast<int>(6* sizeof(float) * vertice.size()), &vertice[0], GL_STATIC_DRAW);
+    glBufferData(GL_ARRAY_BUFFER, static_cast<long long int>(6* sizeof(float) * vertice.size()), &vertice[0], GL_STATIC_DRAW);
 
     glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 6 * sizeof(float), reinterpret_cast<void*>(0));
     glEnableVertexAttribArray(0);
